Move wait_for_checkpoint_start into a checkpoint_util base class
Following the eviction_util/compact_util convention, keep checkpoint test helpers
in their own helpers/checkpoint_util.py rather than the general WiredTigerTestCase
base, so they are easier to find and extend. The layered checkpoint tests now
inherit from checkpoint_util.
